Add config_base.py for Layer
